Move match_path from DefId to lint::LateContext
